backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
description = "Portfolio website backend API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
    "email-validator>=2.2.0",
    "pdfminer.six>=20221105",
    "httpx>=0.27.0",
    "SQLAlchemy>=2.0.35",
    "psycopg[binary]>=3.2.0",
    "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q --tb=short"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]
